Updates to block abstractions
